Add header processing functionality using flag
